+++ sdk/eventhub/azure-eventhub/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 5.6.2 (Unreleased)
+## 5.7.0 (Unreleased)
 
 ### Features Added